User rating Calcukation V1
Creates a separate .csv file for each restaurant. Each csv file has the date and the star rating of a review for the restaurant.

TODO:
Trim the number of restaurants to analyze based on the number of reviews. Preliminary observations suggest that most restaurants have very few reviews, and few having most reviews. No. of reviews v/s No. of restaurants would appear to follow the power law. Plot to confirm.